enigma2 (20120327 rel32 -> 20120430 master) ------------------------------------------- - added default HD skin - added webkit based webbrowser - various changes for hbbtv integration - worked on input device detection - fixed broken live to timeshift transition - fixed transponder scan problems with broken NIT entries (zero symbolrate) - optimized transponder scan (non-DVB transponders are now skipped much faster) - fixed RASS picture download - worked on Common Interface assignment problems and enabled more CI debug output - added option to use custom spinner pictures - fixed teletext subtitle crashes - bugfixes for storage device manager - fixed erroneously removed repeated timers via instant record button - fixed repeated timer handling via instant record button - added notification queue / manager - small skin engine improvements (menu entry description) - added HDMI CEC core support - added hebrew translation - added (expert) menu option to enable SCPC transponder optimized search ranges - fixed slow zapping times between motor-connected LNBs and non-motor LNBs and added a (expert) menu option to switch back to the old behaviour - fixed plugin list oled summary support - added support to burn and playback unprotected "blu discs" - show current ip address in about screen - fixed multi-language epg support (broken since arte/arteHD workaround) - added DLNA Mediaserver - added initial UPnP-Client Support for plugin-developers (enigma2-plugin-systemplugins-upnp), take a look at the available demo plugins for some examples

+++ usr/include/enigma2/lib/dvb/pvrparse.h
@@ -14,16 +14,16 @@ class eMPEGStreamInformation
 public:
 	eMPEGStreamInformation();
 	~eMPEGStreamInformation();
-		/* we order by off_t here, since the timestamp may */
+		/* we order by uint64_t here, since the timestamp may */
 		/* wrap around. */
 		/* we only record sequence start's pts values here. */
-	std::map<off_t, pts_t> m_access_points;
+	std::map<uint64_t, pts_t> m_access_points;
 		/* timestampDelta is in fact the difference between */
 		/* the PTS in the stream and a real PTS from 0..max */
-	std::map<off_t, pts_t> m_timestamp_deltas;
+	std::map<uint64_t, pts_t> m_timestamp_deltas;
 
 		/* these are non-fixed up pts value (like m_access_points), just used to accelerate stuff. */
-	std::multimap<pts_t, off_t> m_pts_to_offset; 
+	std::multimap<pts_t, uint64_t> m_pts_to_offset; 
 
 	int startSave(const char *filename);
 	int stopSave(void);
@@ -33,38 +33,38 @@ public:
 	void fixupDiscontinuties();
 	
 		/* get delta at specific offset */
-	pts_t getDelta(off_t offset);
+	pts_t getDelta(uint64_t offset);
 	
 		/* fixup timestamp near offset, i.e. convert to zero-based */
-	int fixupPTS(const off_t &offset, pts_t &ts);
+	int fixupPTS(const uint64_t &offset, pts_t &ts);
 
 		/* get PTS before offset */	
-	int getPTS(off_t &offset, pts_t &pts);
+	int getPTS(uint64_t &offset, pts_t &pts);
 	
 		/* inter/extrapolate timestamp from offset */
-	pts_t getInterpolated(off_t offset);
+	pts_t getInterpolated(uint64_t offset);
 	
-	off_t getAccessPoint(pts_t ts, int marg=0);
+	uint64_t getAccessPoint(pts_t ts, int marg=0);
 	
 	int getNextAccessPoint(pts_t &ts, const pts_t &start, int direction);
 	
 	bool empty();
 	
-	typedef off_t structure_data;
+	typedef unsigned long long structure_data;
 		/* this is usually:
 			sc | (other_information << 8)
 			but is really specific to the used video encoder.
 		*/
-	void writeStructureEntry(off_t offset, structure_data data);
+	void writeStructureEntry(uint64_t offset, structure_data data);
 
 		/* get a structure entry at given offset (or previous one, if no exact match was found).
 		   optionall, return next element. Offset will be returned. this allows you to easily 
 		   get previous and next structure elements. */
-	int getStructureEntry(off_t &offset, off_t &data, int get_next);
+	int getStructureEntry(uint64_t &offset, uint64_t &data, int get_next);
 
 	std::string m_filename;
 	int m_structure_cache_valid;
-	off_t m_structure_cache[1024];
+	uint64_t m_structure_cache[1024];
 	FILE *m_structure_read, *m_structure_write;
 };
 
@@ -73,14 +73,14 @@ class eMPEGStreamParserTS
 {
 public:
 	eMPEGStreamParserTS(eMPEGStreamInformation &streaminfo);
-	void parseData(off_t offset, const void *data, unsigned int len);
+	void parseData(uint64_t offset, const void *data, unsigned int len);
 	void setPid(int pid, int streamtype);
 	int getLastPTS(pts_t &last_pts);
 private:
 	eMPEGStreamInformation &m_streaminfo;
 	unsigned char m_pkt[188];
 	int m_pktptr;
-	int processPacket(const unsigned char *pkt, off_t offset);
+	int processPacket(const unsigned char *pkt, uint64_t offset);
 	inline int wantPacket(const unsigned char *hdr) const;
 	int m_pid, m_streamtype;
 	int m_need_next_packet;
